+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## [Unreleased][unreleased]
 
+## [2.0.11][] - 2021-02-28
+
 - Schema for collections: array, object, set, map
+- Change queue configuration: https://github.com/metarhia/impress/issues/1484
 
 ## [2.0.10][] - 2021-02-24
 
@@ -95,7 +98,8 @@ First generation of application server with following features
 - Connection drivers for database engines: MongoDB, PgSQL, Oracle, MySQL
 - Support GeoIP, health monitoring, task scheduling, server-side templating
 
-[unreleased]: https://github.com/metarhia/impress/compare/v2.0.10...HEAD
+[unreleased]: https://github.com/metarhia/impress/compare/v2.0.11...HEAD
+[2.0.11]: https://github.com/metarhia/impress/compare/v2.0.10...v2.0.11
 [2.0.10]: https://github.com/metarhia/impress/compare/v2.0.9...v2.0.10
 [2.0.9]: https://github.com/metarhia/impress/compare/v2.0.8...v2.0.9
 [2.0.8]: https://github.com/metarhia/impress/compare/v2.0.7...v2.0.8